Zak's Final Submission to E.INC
May 10 2022
I wanted to make my submission at least functional regardless of being incomplete. In the final hour of the project, some of the more complex design elements had to be scrapped. For instance, the opening HERO area: I wanted to perform some animation effects on the 3 images but ran out of time and ended up hiding them.

What I'm proud of:
- the site is functional
- fully responsive (although, with more time the responsivity experience definitely could be improved
- The FORM design looks great and is responsive (mostly...would have improved it at 440px and below)
- sourced and used Edealer fonts

What I'm ashamed of:
- no icons
- the 'get started' banner doesn't have a ribbon nor box shadow
- the HERO area
- the are some responsive issues with large spaces and/or font-sizes that could've been improved